e2e test feature
