+++ sf_core/tests/e2e/mod.rs
@@ -3,3 +3,4 @@ pub mod common;
 
 pub mod authentication;
 pub mod query;
+pub mod tls;
