--- sf_core/tests/e2e_tls_tests.rs
@@ -1,35 +1,0 @@
-#[cfg(feature = "e2e_tests")]
-mod e2e {
-    use sf_core::tls::config::TlsConfig;
-    use sf_core::tls::create_tls_client_with_config;
-
-    #[tokio::test]
-    async fn handshake_with_default_roots() {
-        let server_url =
-            std::env::var("E2E_TLS_SERVER").unwrap_or("https://example.com".to_string());
-        let client = create_tls_client_with_config(TlsConfig::default()).expect("client");
-        let resp = client.get(server_url).send().await;
-        // Only assert we can attempt; don't fail CI when network blocked
-        assert!(resp.is_ok() || resp.is_err());
-    }
-
-    #[tokio::test]
-    async fn handshake_with_custom_roots_env() {
-        if let Ok(pem_path) = std::env::var("E2E_TLS_ROOTS_PEM") {
-            let cfg = TlsConfig {
-                custom_root_store_path: Some(pem_path.into()),
-                ..Default::default()
-            };
-            let client = create_tls_client_with_config(cfg).expect("client");
-            let server_url =
-                std::env::var("E2E_TLS_SERVER").unwrap_or("https://example.com".to_string());
-            let _ = client.get(server_url).send().await;
-        }
-    }
-}
-
-#[cfg(not(feature = "e2e_tests"))]
-#[test]
-fn e2e_tests_feature_gated() {
-    // Intentionally empty: ensures test file compiles without running external calls
-}
